Fixed the NPE when looking up the next test class
